fix: Resolve all Clippy errors and warnings
- Fixed type conversion issues between u64 and cryptoki::types::Ulong
- Removed unused imports and variables
- Added clippy allow directives for acceptable patterns
- Fixed all compilation errors from strict Clippy checks

+++ crates/rust-hsm-cli/src/pkcs11/keys/keypair.rs
@@ -54,7 +54,7 @@ pub fn gen_keypair(
                 Attribute::Id(key_label.as_bytes().to_vec()),
                 Attribute::Encrypt(true),
                 Attribute::Verify(true),
-                Attribute::ModulusBits(cryptoki::types::Ulong::from(bits as u64)),
+                Attribute::ModulusBits(cryptoki::types::Ulong::from(bits)),
                 Attribute::PublicExponent(vec![0x01, 0x00, 0x01]), // 65537
             ];
